feat: 캐시 프록시 패턴 구현
