+++ proxy/src/test/java/com/example/proxy/pureproxy/proxy/code/ProxyPatternClient.java
@@ -1,4 +1,4 @@
-package com.example.proxy.code;
+package com.example.proxy.pureproxy.proxy.code;
 
 public class ProxyPatternClient {
