+++ proxy/src/test/java/com/example/proxy/pureproxy/proxy/code/Subject.java
@@ -1,4 +1,4 @@
-package com.example.proxy.code;
+package com.example.proxy.pureproxy.proxy.code;
 
 public interface Subject {
     String operation();
